backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pymysql==1.1.0
cryptography==41.0.7
python-multipart==0.0.6
python-dotenv==1.0.0
dashscope>=1.10.0
Pillow==10.1.0
aiofiles==23.2.1
requests>=2.31.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

